Revert "Test: Add an integration test in which one N parties behave maliciously" (#1387)
