+++ crates/ika-core/src/dwallet_mpc/integration_tests/mod.rs
@@ -1,4 +1,3 @@
-mod malicious_behavior;
 #[cfg(test)]
 mod network_dkg;
 mod utils;
